pythonapp/requirements.txt
protobuf
mss
pillow==7.1.2
numpy
PySide2
d3dshot
py2app